refac: error 정리하기
* error 정리하기
  - useError에서는 UI에 표시되는 vaildation 메세지 관련해서 관리하고 사용하는 쪽으로 정의
  - 기존에 alertError 로직들은 error.ts에 정의해서 중복적으로 사용하는 곳에서 import
  해와서 함수처럼 사용..

2023.08.10
