+++ src/context/AuthContext.tsx
@@ -1,13 +1,10 @@
-import { ReactNode, createContext, useState, useEffect } from 'react'
+import { ReactNode, createContext, useState } from 'react'
 import localToken from '../api/LocalToken'
 
 export interface AuthContextProps {
   email: string
   password: string
-  btnDisabled: boolean
   accessToken: string | null
-  isValidEmail: boolean
-  isValidPassword: boolean
   handleEmailChange: (e: React.ChangeEvent<HTMLInputElement>) => void
   handlePasswordChange: (e: React.ChangeEvent<HTMLInputElement>) => void
 }
@@ -17,16 +14,8 @@ export const AuthContext = createContext<AuthContextProps | null>(null)
 export const AuthContextProvider = ({ children }: { children: ReactNode }) => {
   const [email, setEmail] = useState('')
   const [password, setPassword] = useState('')
-  const [btnDisabled, setBtnDisabled] = useState(false)
 
   const accessToken = localToken.get()
-
-  const isValidEmail = email.indexOf('@') !== -1
-  const isValidPassword = password.length >= 8
-
-  useEffect(() => {
-    setBtnDisabled(!isValidEmail || !isValidPassword)
-  }, [isValidEmail, isValidPassword])
 
   const handleEmailChange = (e: React.ChangeEvent<HTMLInputElement>): void => {
     setEmail(e.target.value)
@@ -41,10 +30,7 @@ export const AuthContextProvider = ({ children }: { children: ReactNode }) => {
   const contextValue: AuthContextProps = {
     email,
     password,
-    btnDisabled,
     accessToken,
-    isValidEmail,
-    isValidPassword,
     handleEmailChange,
     handlePasswordChange
   }
