+++ src/styles/TodoStyle.tsx
@@ -83,7 +83,7 @@ export const Label = styled.label<{ completed: boolean }>`
   overflow-x: auto;
   margin-left: 20px;
   width: 200px;
-  height: 20px;
+  height: 21px;
   text-decoration: ${(props) => (props.completed ? 'line-through' : 'none')};
 `
