+++ src/utils/message.ts
@@ -3,5 +3,8 @@ export const UNKNOWN_ERROR = '알 수 없는 에러가 발생했습니다.'
 export const VALID_MESSAGE_EMAIL = '이메일에 @ 기호가 없습니다.'
 export const VALID_MESSAGE_PASSWORD = '비밀번호는 8자 이상이여야 합니다.'
 
+export const CHECK_TODO_INPUT =
+  '내용을 입력한 후 추가 버튼이나 Enter를 눌러주세요.'
+
 export const COMPLETED_SIGN_UP = '회원가입이 완료되었습니다.'
 export const COMPLETED_SIGN_IN = '로그인이 완료되었습니다!'
